requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
aiofiles
torch
transformers
Pillow
matplotlib
opencv-python
